added functions in sql write
